extend utils to everything
