+++ frontend/app/novels/novels-stats/page.tsx
@@ -2,7 +2,7 @@
 
 import { Stats } from "./stats.tsx";
 import React, {useState, useEffect} from 'react';
-import { BackendRequest, fetch_backend } from "@/utils/fetch_backend.ts";
+import { fetch_backend } from "@/utils/fetch_backend.ts";
 
 export default function Page() {
   const [stats, setStats] = useState<Stats>();
@@ -10,7 +10,7 @@ export default function Page() {
   // load novel stats once
   useEffect(() => {
     const fetchStats = async () => {
-      const data = await fetch_backend({path: "/api/novels_stats", method: "GET", body: null});
+      const data = await fetch_backend({path: "/api/novels_stats", method: "GET", body: undefined});
       setStats(data);
     };
